Add `cut` expressions
A cut expressions forces the matcher to only consider the current path of an alternative expression. If a cut expression is reached when trying to match one of many options inside an alternative expression, no other ones will be considered after. In case of a match failure of this option the whole alternative expression fails.